initial commit for dynamic URL loading support it supports both childFirst and parentFirst  jarLoading added basic test coverage
