+++ cluster/src/test/scala/org/apache/spark/DynamicURLClassLoaderSuite.scala
@@ -18,9 +18,14 @@
 
 package org.apache.spark
 import java.net.URLClassLoader
+
 import scala.collection.JavaConverters._
+
+import com.gemstone.gemfire.cache.CacheClosedException
 import org.scalatest.Matchers
-import org.apache.spark.util.{DynamicURLClassLoader, ChildFirstURLClassLoader, MutableURLClassLoader, Utils}
+
+import org.apache.spark.util.Utils
+import org.apache.spark.util.classloader.DynamicURLClassLoader
 
 
 class DynamicURLClassLoaderSuite extends SparkFunSuite with Matchers {
@@ -40,34 +45,21 @@ class DynamicURLClassLoaderSuite extends SparkFunSuite with Matchers {
   val fileUrlsParent = List(TestUtils.createJarWithFiles(Map(
     "resource1" -> "resource1Contents-parent"))).toArray
 
+
   test("child first") {
-    val parentLoader = new URLClassLoader(urls2, null)
-    val classLoader = new DynamicURLClassLoader(urls, parentLoader, true)
+    val parentLoader = new URLClassLoader(urls2, Utils.getContextOrSparkClassLoader)
+    val classLoader = new DynamicURLClassLoader(urls, parentLoader, parentFirst = false)
     val fakeClass = classLoader.loadClass("FakeClass2").newInstance()
     val fakeClassVersion = fakeClass.toString
-    //assert(fakeClassVersion === "2")
+    assert(fakeClassVersion === "1")
     val fakeClass2 = classLoader.loadClass("FakeClass2").newInstance()
-    assert(fakeClass.getClass == fakeClass2.getClass)
-    assert(fakeClass.getClass.getClassLoader == fakeClass2.getClass.getClassLoader)
-    println ("loader is " + fakeClass.getClass.getClassLoader.getClass.getName)
+    assert(fakeClass.getClass === fakeClass2.getClass)
   }
 
-  test("child first when jars are added at run time") {
-    val parentLoader = getClass.getClassLoader
-    val classLoader = new DynamicURLClassLoader(urls2, parentLoader, true)
-    urls.foreach(classLoader.addURL)
-    val fakeClass = classLoader.loadClass("FakeClass2").newInstance()
-    val fakeClassVersion = fakeClass.toString
-//    assert(fakeClassVersion === "2")
-    val fakeClass2 = classLoader.loadClass("FakeClass2").newInstance()
-    assert(fakeClass.getClass == fakeClass2.getClass)
-    assert(fakeClass.getClass.getClassLoader == fakeClass2.getClass.getClassLoader)
-    println ("loader is " + fakeClass.getClass.getClassLoader.getClass.getName)
-  }
 
   test("parent first") {
-    val parentLoader = new URLClassLoader(urls2, null)
-    val classLoader = new MutableURLClassLoader(urls, parentLoader)
+    val parentLoader = new URLClassLoader(urls2, Utils.getContextOrSparkClassLoader)
+    val classLoader = new DynamicURLClassLoader(urls, parentLoader, parentFirst = true)
     val fakeClass = classLoader.loadClass("FakeClass1").newInstance()
     val fakeClassVersion = fakeClass.toString
     assert(fakeClassVersion === "2")
@@ -76,8 +68,8 @@ class DynamicURLClassLoaderSuite extends SparkFunSuite with Matchers {
   }
 
   test("child first can fall back") {
-    val parentLoader = new URLClassLoader(urls2, null)
-    val classLoader = new ChildFirstURLClassLoader(urls, parentLoader)
+    val parentLoader = new URLClassLoader(urls2, Utils.getContextOrSparkClassLoader)
+    val classLoader = new DynamicURLClassLoader(urls, parentLoader, parentFirst = false)
     val fakeClass = classLoader.loadClass("FakeClass3").newInstance()
     val fakeClassVersion = fakeClass.toString
     assert(fakeClassVersion === "2")
@@ -85,29 +77,30 @@ class DynamicURLClassLoaderSuite extends SparkFunSuite with Matchers {
 
   test("child first can fail") {
     val parentLoader = new URLClassLoader(urls2, null)
-    val classLoader = new ChildFirstURLClassLoader(urls, parentLoader)
-    intercept[java.lang.ClassNotFoundException] {
+    val classLoader =
+      new DynamicURLClassLoader(urls, parentLoader, parentFirst = true)
+    intercept[CacheClosedException] {
       classLoader.loadClass("FakeClassDoesNotExist").newInstance()
     }
   }
 
   test("default JDK classloader get resources") {
     val parentLoader = new URLClassLoader(fileUrlsParent, null)
-    val classLoader = new URLClassLoader(fileUrlsChild, parentLoader)
+    val classLoader = new DynamicURLClassLoader(fileUrlsChild, parentLoader, false)
     assert(classLoader.getResources("resource1").asScala.size === 2)
     assert(classLoader.getResources("resource2").asScala.size === 1)
   }
 
   test("parent first get resources") {
     val parentLoader = new URLClassLoader(fileUrlsParent, null)
-    val classLoader = new MutableURLClassLoader(fileUrlsChild, parentLoader)
+    val classLoader = new DynamicURLClassLoader(fileUrlsChild, parentLoader, parentFirst = true)
     assert(classLoader.getResources("resource1").asScala.size === 2)
     assert(classLoader.getResources("resource2").asScala.size === 1)
   }
 
   test("child first get resources") {
     val parentLoader = new URLClassLoader(fileUrlsParent, null)
-    val classLoader = new ChildFirstURLClassLoader(fileUrlsChild, parentLoader)
+    val classLoader = new DynamicURLClassLoader(fileUrlsChild, parentLoader, parentFirst = false)
 
     val res1 = classLoader.getResources("resource1").asScala.toList
     assert(res1.size === 2)
@@ -115,38 +108,5 @@ class DynamicURLClassLoaderSuite extends SparkFunSuite with Matchers {
 
     res1.map(scala.io.Source.fromURL(_).mkString) should contain inOrderOnly
         ("resource1Contents-child", "resource1Contents-parent")
-  }
-
-
-  test("driver sets context class loader in local mode") {
-    // Test the case where the driver program sets a context classloader and then runs a job
-    // in local mode. This is what happens when ./spark-submit is called with "local" as the
-    // master.
-    val original = Thread.currentThread().getContextClassLoader
-
-    val className = "ClassForDriverTest"
-    val jar = TestUtils.createJarWithClasses(Seq(className))
-    val contextLoader = new URLClassLoader(Array(jar), Utils.getContextOrSparkClassLoader)
-    Thread.currentThread().setContextClassLoader(contextLoader)
-
-    val sc = new SparkContext("local", "driverLoaderTest")
-
-    try {
-      sc.makeRDD(1 to 5, 2).mapPartitions { x =>
-        val loader = Thread.currentThread().getContextClassLoader
-        // scalastyle:off classforname
-        Class.forName(className, true, loader).newInstance()
-        // scalastyle:on classforname
-        Seq().iterator
-      }.count()
-    }
-    catch {
-      case e: SparkException if e.getMessage.contains("ClassNotFoundException") =>
-        fail("Local executor could not find class", e)
-      case t: Throwable => fail("Unexpected exception ", t)
-    }
-
-    sc.stop()
-    Thread.currentThread().setContextClassLoader(original)
   }
 }
